backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
